feat: add `reset-timeout-on-press` to `tap-hold-release-timeout` (#1696)
Add optional flag argument `reset-timeout-on-press` for
`tap-hold-release-timeout`, which if included changes the timeout
behaviour. This flag will make the timeout duration reset if a new key
is pressed. This may result in longer timeouts, but can help with more
consistent hold activations, because it may be challenging to release a
key in time to activate the hold action instead of the timeout action.