fix: deallocate correct size
This commit represents allocations as a Box<[u8]> so that dropping the
Box frees the correct size. This stops dlmalloc from panicking on
formerly incorrect unsafe code.

Likely the "proper" way to do this would be using currently-unstable
alloc module and saving layout information alongside the pointer, and
then dealloc can be called directly.