feat: consume driverkit-v0.2.0 (#1813)
Changes in driverkit v0.2.0:
1. Major re-architecture. Simpler and more linear flow.
2. Provide new interface to query driverkit for devices
`fetch_devices()`
3. Expose a hash value for devices that can be used to register them
with Kanata -> enables using keyboards without a name (#1668)
4. Fix issues with tracking devices when they're disconnected -> now
should always re-capture the device when it re-connects (given it was
registered to start with).

Changes in Kanata:
1. Remove `capture_stdout()` and replace it with
`driverkit::fetch_devices()` -> a much better solution.
2. Kanata -l (--list) now provides more comprehensive information,
including device hashes -> new feature in driverkit
3. kanata can now register any device with it's hash instead of it's
name, hashes and names are now usable in the include and exclude lists.

This change resolves: #1787, #1668 and probably #1479 and #1539 too
I couldn't test the last two mentioned issues because I don't have the
hardware to do so.
Would be awesome if someone can confirm if they were indeed fixed or
not.

Co-authored-by: Abdalrahman Mursi <amursi@mentor.com>