fix: cross compilation for windows (#1900)
Cross compilation for Windows is currently not possible because
`target_os` deviates from its usual meaning when used for
`build-dependencies` and inside `build.rs`. There, it refers to the
`host` instead of the `target`.
As the `build-dependencies` are already marked optional and only enabled
through the `win_manifest` feature, this PR simply removes the check for
the `target_os`.
There does not seem to be a config based mechanism to only allow the
`win_manifest` feature on a Windows target. If desired, such a check
could however be included in the `build.rs` as a "runtime" check, i.e.,
at the time the buildscript is executed.

Additionally, this PR hides the features automatically generated for
optional dependencies.

Best,
PhiPro